+++ src/hooks/useBlockingQuery.ts
@@ -37,6 +37,8 @@ export function useBlockingQuery<
     reasonOnFetching,
     reasonOnError,
     priority = 10,
+    timeout,
+    onTimeout,
     onLoading = true,
     onFetching = false,
     onError = false,
@@ -63,6 +65,8 @@ export function useBlockingQuery<
       scope,
       reason: currentReason,
       priority,
+      timeout,
+      onTimeout,
     },
     [
       query.isLoading,